+++ packages/app-page-builder/src/admin/views/Pages/Pages.tsx
@@ -35,6 +35,7 @@ const Pages = () => {
             if (error) {
                 showSnackbar(error.message);
             } else {
+                console.log("[Pages]");
                 history.push(`/page-builder/editor/${encodeURIComponent(data.id)}`);
             }
         } catch (e) {